# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_jcom-dev_zmanim_internal_models.City:
    properties:
      continent:
        description: Continent code (AF, AN, AS, EU, NA, OC, SA)
        type: string
      country:
        type: string
      country_code:
        type: string
      display_name:
        description: Computed display field
        type: string
      elevation:
        description: Elevation in meters above sea level
        type: integer
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      population:
        type: integer
      region:
        type: string
      region_type:
        type: string
      timezone:
        type: string
    type: object
  github_com_jcom-dev_zmanim_internal_models.CitySearchResponse:
    properties:
      cities:
        items:
          $ref: '#/definitions/github_com_jcom-dev_zmanim_internal_models.City'
        type: array
      total:
        type: integer
    type: object
...
  internal_handlers.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  internal_handlers.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/internal_handlers.APIError'
      meta:
        $ref: '#/definitions/internal_handlers.ResponseMeta'
    type: object
...
  internal_handlers.ResponseMeta:
    properties:
      request_id:
        type: string
      timestamp:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@zmanim.com
    name: Zmanim Lab Support
  description: API for Jewish prayer times calculation and publisher management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://zmanim.com/terms
  title: Zmanim Lab API
  version: "1.0"
paths:
...
  /cities:
    get:
      description: Search for cities by name with optional filtering by country, region,
        or continent. Uses fuzzy matching for typo tolerance.
      parameters:
      - description: Search query (min 2 chars for fuzzy match)
        in: query
        name: search
        type: string
      - description: ISO 3166-1 alpha-2 country code
        in: query
        name: country_code
        type: string
      - description: Region/state name
        in: query
        name: region
        type: string
      - description: Continent code (AF, AS, EU, NA, OC, SA, AN)
        in: query
        name: continent_code
        type: string
      - description: Max results (default 10, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of matching cities
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_jcom-dev_zmanim_internal_models.CitySearchResponse'
              type: object
        "400":
          description: Invalid parameters
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
      summary: Search cities
      tags:
      - Cities
...
  /cities/nearby:
    get:
      description: Finds the nearest city to the given GPS coordinates using PostGIS
        spatial queries
      parameters:
      - description: Latitude (-90 to 90)
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude (-180 to 180)
        in: query
        name: lng
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: Nearest city with distance
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Invalid coordinates
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
      summary: Find nearest city
      tags:
      - Cities
  /continents:
    get:
      description: Returns all continents with their city counts
      produces:
      - application/json
      responses:
        "200":
          description: List of continents
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
      summary: List continents
      tags:
      - Geography
  /countries:
    get:
      description: Returns all countries with their city counts, optionally filtered
        by continent
      parameters:
      - description: Filter by continent code or name
        in: query
        name: continent
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of countries
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handlers.APIError'
              type: object
      summary: List countries
      tags:
      - Geography
...
securityDefinitions:
  BearerAuth:
    description: JWT Bearer token from Clerk authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"